making parent protocol class
