+++ qadence_protocols/mitigations/protocols.py
@@ -5,6 +5,8 @@ from dataclasses import dataclass
 from functools import partial
 from typing import Callable
 
+from qadence_protocols.protocols import Protocol
+
 PROTOCOL_TO_MODULE = {
     "twirl": "qadence_protocols.mitigations.twirl",
     "readout": "qadence_protocols.mitigations.readout",
@@ -13,14 +15,13 @@ PROTOCOL_TO_MODULE = {
 
 
 @dataclass
-class Mitigations:
+class Mitigations(Protocol):
     TWIRL = "twirl"
     READOUT = "readout"
     ANALOG_ZNE = "zne"
 
     def __init__(self, protocol: str, options: dict = dict()) -> None:
-        self.protocol: str = protocol
-        self.options: dict = options
+        super().__init__(protocol, options)
 
     def mitigation(self) -> Callable:
         try:
@@ -29,16 +30,3 @@ class Mitigations:
             ImportError(f"The module for the protocol {self.protocol} is not implemented.")
         # Partially pass the options.
         return partial(getattr(module, "mitigate"), options=self.options)
-
-    def _to_dict(self) -> dict:
-        return {"protocol": self.protocol, "options": self.options}
-
-    @classmethod
-    def _from_dict(cls, d: dict) -> Mitigations | None:
-        if d:
-            return cls(d["protocol"], **d["options"])
-        return None
-
-    @classmethod
-    def list(cls) -> list:
-        return list(filter(lambda el: not el.startswith("__"), dir(cls)))
