+++ qadence_protocols/protocols.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import importlib
+from typing import Any
 
 from qadence_protocols.types import qadence_available_protocols
 
@@ -16,3 +17,24 @@ def available_protocols() -> dict:
         qadence_protocols[protocol] = ProtocolCls
 
     return qadence_protocols
+
+
+class Protocol:
+    """Generic class for protocols."""
+
+    def __init__(self, protocol: str, options: dict = dict()) -> None:
+        self.protocol: str = protocol
+        self.options: dict = options
+
+    def _to_dict(self) -> dict:
+        return {"protocol": self.protocol, "options": self.options}
+
+    @classmethod
+    def _from_dict(cls, d: dict) -> Any | None:
+        if d:
+            return cls(d["protocol"], **d["options"])
+        return None
+
+    @classmethod
+    def list(cls) -> list:
+        return list(filter(lambda el: not el.startswith("__"), dir(cls)))
